put sdf shit on stack

+++ include/physics/physics.h
@@ -17,7 +17,7 @@ typedef struct srph_physics {
 	  std::mutex substances_mutex;
 
 	substance_t *substances;
-	size_t *num_substances;
+	uint32_t *num_substances;
 
 	srph_collision_array collisions;
 
@@ -27,7 +27,7 @@ typedef struct srph_physics {
 } srph_physics;
 
 void srph_physics_init(srph_physics * p, substance_t * substances,
-	size_t *num_substances);
+                       uint32_t *num_substances);
 void srph_physics_start(srph_physics * p);
 void srph_physics_destroy(srph_physics * p);
 void srph_physics_tick(srph_physics * p, double dt);
